requirements.txt
Django==5.1.6
djangorestframework==3.15.2
asgiref==3.8.1
boto3==1.36.26
botocore==1.36.26
django-webpack-loader==3.1.1
django-cors-headers==4.7.0
s3transfer==0.11.2
django-environ==0.12.0
django-filter==25.1
django-axes==7.0.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-storages==1.14.5
django-ses==4.1.0
dj-rest-auth==7.0.1
django-allauth==65.4.1
django-auto-logout==0.5.1
requests==2.32.2
pandas==2.2.3
django-csp==4.0b4